refactor: remove retired feature screens
