+++ shared/src/commonMain/kotlin/org/feeluown/mobile/feature/settings/SettingsUiPrimitives.kt
@@ -1,0 +1,27 @@
+package org.feeluown.mobile
+
+import androidx.compose.material3.FilterChipDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.SegmentedButtonDefaults
+import androidx.compose.runtime.Composable
+
+const val FUO_EVOLVE_SOURCE_URL = "https://github.com/feeluown/FuoEvolve"
+const val FEELUOWN_SOURCE_URL = "https://github.com/feeluown/FeelUOwn"
+
+@Composable
+fun settingsSegmentedButtonColors() = SegmentedButtonDefaults.colors(
+    activeContainerColor = MaterialTheme.colorScheme.primary,
+    activeContentColor = MaterialTheme.colorScheme.onPrimary,
+    activeBorderColor = MaterialTheme.colorScheme.primary,
+    inactiveContainerColor = MaterialTheme.colorScheme.surface,
+    inactiveContentColor = MaterialTheme.colorScheme.onSurface,
+    inactiveBorderColor = MaterialTheme.colorScheme.outline,
+)
+
+@Composable
+fun settingsFilterChipColors() = FilterChipDefaults.filterChipColors(
+    selectedContainerColor = MaterialTheme.colorScheme.primary,
+    selectedLabelColor = MaterialTheme.colorScheme.onPrimary,
+    containerColor = MaterialTheme.colorScheme.surface,
+    labelColor = MaterialTheme.colorScheme.onSurface,
+)
